# Simplify schema attributes and consolidate definitions