--- a/src/DocumentFormat.OpenXml/Cached.cs
+++ b/src/DocumentFormat.OpenXml/Cached.cs
@@ -1,7 +1,9 @@
 ﻿// Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
+using System;
 using System.Collections.ObjectModel;
+using System.Linq.Expressions;
 
 namespace DocumentFormat.OpenXml
 {
@@ -23,6 +25,27 @@ namespace DocumentFormat.OpenXml
         private static class ReadOnlyCollectionCache<T>
         {
             internal static ReadOnlyCollection<T> Value = new ReadOnlyCollection<T>(Array<T>());
+        }
+
+        public static Func<T> Activator<T>() where T : new() => CachedActivator<T>.Activator;
+
+        private static class CachedActivator<T>
+            where T : new()
+        {
+            private static readonly Expression<Func<T>> Creator = () => new T();
+
+            public static Func<T> Activator { get; } = Creator.Compile();
+        }
+
+        public static Func<TBase> Activator<TInstance, TBase>()
+            where TInstance : TBase, new() => CachedActivator<TInstance, TBase>.Activator;
+
+        private static class CachedActivator<TInstance, TBase>
+            where TInstance : TBase, new()
+        {
+            private static readonly Expression<Func<TBase>> Creator = () => new TInstance();
+
+            public static Func<TBase> Activator { get; } = Creator.Compile();
         }
     }
 }