--- a/src/DocumentFormat.OpenXml/GeneratedCode/schemas-microsoft-com_office_powerpoint.g.cs
+++ b/src/DocumentFormat.OpenXml/GeneratedCode/schemas-microsoft-com_office_powerpoint.g.cs
@@ -71,7 +71,7 @@ public partial class TextData : OpenXmlLeafElement
 	{
 		AttributeTag.Create<StringValue>(0, "id")
 	};
-    internal override AttributeTagCollection RawAttributes { get; } = s_attributeTags;
+    internal override AttributeTagCollection RawAttributes { get; } = new AttributeTagCollection(s_attributeTags);
 
     
         /// <summary>