--- a/src/DocumentFormat.OpenXml/GeneratedCode/schemas_microsoft_com_office_2006_metadata_customXsn.g.cs
+++ b/src/DocumentFormat.OpenXml/GeneratedCode/schemas_microsoft_com_office_2006_metadata_customXsn.g.cs
@@ -155,7 +155,6 @@ if( 39 == namespaceId && "xsnScope" == name)
     }
 
 
-    
     /// <inheritdoc/>
     public override OpenXmlElement CloneNode(bool deep) => CloneImp<CustomXsn>(deep);