This creates a readonly struct that abstracts the attribute definitions
and management to help with readability and cleaner code.

Before:

```
BenchmarkDotNet=v0.10.14, OS=Windows 10.0.14393.2155 (1607/AnniversaryUpdate/Redstone1), VM=Hyper-V
Intel Xeon CPU E5-2673 v4 2.30GHz, 1 CPU, 2 logical and 2 physical cores
  [Host] : .NET Framework 4.7.1 (CLR 4.0.30319.42000), 64bit RyuJIT-v4.7.2558.0
  net46  : .NET Framework 4.7.1 (CLR 4.0.30319.42000), 64bit RyuJIT-v4.7.2558.0

Job=net46  Runtime=Clr  

|      Method|      Mean|     Error|    StdDev|      Gen 0|     Gen 1|     Gen 2|  Allocated
|  Validation|  79.86 ms|  1.519 ms|  1.560 ms|  2437.5000|  687.5000|  312.5000|   13.86 MB
```

After:

```
BenchmarkDotNet=v0.10.14, OS=Windows 10.0.14393.2155 (1607/AnniversaryUpdate/Redstone1), VM=Hyper-V
Intel Xeon CPU E5-2673 v4 2.30GHz, 1 CPU, 2 logical and 2 physical cores
  [Host] : .NET Framework 4.7.1 (CLR 4.0.30319.42000), 64bit RyuJIT-v4.7.2558.0
  net46  : .NET Framework 4.7.1 (CLR 4.0.30319.42000), 64bit RyuJIT-v4.7.2558.0

Job=net46  Runtime=Clr  

|      Method|      Mean|     Error|    StdDev|      Gen 0|     Gen 1|     Gen 2|  Allocated
|  Validation|  80.01 ms|  1.535 ms|  1.198 ms|  2500.0000|  687.5000|  187.5000|   13.91 MB
```

There is a large reduction of Gen2 allocations, while everything else remains similar. This change reduces the code to handle attributes from 60k LOC to 30k, reduces 3 virtual methods to a single one per schema, and reduces the static arrays per schema from 2 to 1.  A future request will attempt to bring a similar system to the element definitions/creation.